refactor(mcp-server): let the document store stop knowing about compaction
`document-store.ts` held two things. One is what its name says: reading and
writing documents. The other is a background scheduler — per-document debounce
timers, an in-flight promise set, a disposal protocol with its own counter, and
two test-only introspection hooks. It moves to `auto-compact.ts`.

The move alone does not work, and finding out why is the actual change.
Lifting the block left `saveDocument` calling `autoCompactTrigger`, a variable
that had gone with it: the dependency ran BOTH ways. The store called the
scheduler and the scheduler called the store, which is why the two had never
come apart.

The seam was misnamed rather than misplaced. What `saveDocument` has to say is
"this document changed"; who cares is not its business. So the store keeps a
plain `setDocumentSavedListener`, knowing nothing about compaction, and
`installAutoCompact(versionStore)` subscribes. The version-store type stays out
of the store exactly as before — that indirection was the original reason for a
registered trigger and it survives intact.

`setAutoCompactTrigger(null)` is now `uninstallAutoCompact()`, which is what it
always did: drop the listener AND clear the timers. Its old name described the
first half and its behaviour did both.

The DB dispose hook moves with the scheduler. It used to register whenever
`document-store.ts` loaded — early and often — and now registers when
`auto-compact.ts` does, which is exactly when a compaction first becomes
possible, since nothing can schedule one without importing that file.

Behaviour-preserving, and checked as such: 112 tests over `document-store`
before, 112 after, no test weakened, and arch-lint's cycle check passes on the
new one-way edge. 928 lines becomes 802 + 178.

One test was renamed to describe the seam rather than its old subscriber —
`saveDocument invokes the registered auto-compact trigger` is now
`saveDocument notifies the registered document-saved listener`.

Splitting the 2155-line test file along the same line is a separate increment;
this one moves no tests, so a case cannot go quiet in the move.
